+++ AppointmentAPI/Services/Interface/IService.cs
@@ -5,11 +5,11 @@
 
         IEnumerable<T> GetAll();
 
-        T Get(int id);
+        T Get(long id);
 
         void Update(T entity);
 
-        void Delete(int id);
+        void Delete(long id);
 
         int SaveChanges();
